==> vmags_check_helper/01_keyword_check.ipynb <==
import os
import pandas as pd
import numpy as np
import glob

amg_contigs_for_checking = [f.replace('__', '||') for f in np.loadtxt('/fs/ess/PAS0439/MING/virome/amg_analysis/amg_containing_contigs_curated_vs2.txt', dtype = 'str')]

amr_contigs_for_checking = np.loadtxt('/fs/ess/PAS0439/MING/virome/AMR_AMG_containing_contigs_manual_check/amr_containing_contigs.txt', dtype = 'str')

os.chdir('/fs/ess/PAS0439/MING/virome/checkv_res')

checkv_files = glob.glob('**/quality_summary.tsv', recursive = True)

checkv_df = []
for file in checkv_files:
    df = pd.read_csv(file, sep = '\t')
    checkv_df.append(df)

checkv_for_manual_curation = pd.concat(checkv_df)

checkv_for_amg_curation = checkv_for_manual_curation[checkv_for_manual_curation.contig_id.isin(amg_contigs_for_checking)][['contig_id', 'contig_length', 'gene_count', 'viral_genes', 'host_genes', 'checkv_quality']].reset_index(drop = True)

checkv_for_amg_curation

checkv_for_amr_curation = checkv_for_manual_curation[checkv_for_manual_curation.contig_id.isin(amr_contigs_for_checking)][['contig_id', 'contig_length', 'gene_count','viral_genes','host_genes', 'checkv_quality']].reset_index(drop = True)

#pd.set_option('display.max_rows', None)
checkv_for_amr_curation.query('~(host_genes > 0 & viral_genes == 0)').query('contig_length > 10000')

os.chdir('/fs/ess/PAS0439/MING/virome/amg_analysis/dram_annotation')

dramv_amg = glob.glob('**/annotations.tsv', recursive = True)

dramv_amg_df = []
for file in dramv_amg:
    df = pd.read_csv(file, sep = '\t')
    dramv_amg_df.append(df)

dramv_amg_for_manual_curation = pd.concat(dramv_amg_df)

dramv_amg_for_manual_curation.rename(columns = {'Unnamed: 0': 'gene_name'}, inplace = True)
dramv_amg_for_manual_curation.fasta = dramv_amg_for_manual_curation.fasta.apply(lambda x: x.replace('__full-cat_1', '||full'))

dramv_amg_for_manual_curation_reformat = dramv_amg_for_manual_curation.get(['gene_name', 'fasta', 'kegg_hit', 'viral_hit', 'pfam_hits'])
dramv_amg_for_manual_curation_reformat_new = dramv_amg_for_manual_curation_reformat[dramv_amg_for_manual_curation_reformat.fasta.isin(amg_contigs_for_checking)]

group = dramv_amg_for_manual_curation_reformat_new.groupby('fasta').groups.keys()

import re
def find_text(ser, string):
    res = ser[ser.str.contains(string, flags=re.IGNORECASE)]
    if res.empty:
        res = ''
    else:
        res = str(list(res))
    return res
def get_hallmark(df):
    kegg_hit = df['kegg_hit'].fillna('')
    viral_hit = df['viral_hit'].fillna('')
    pfam_hit =  df['pfam_hits'].fillna('')
    
       
    kegg_capsid = find_text(kegg_hit, 'capsid')
    kegg_portal = find_text(kegg_hit, 'portal')
    kegg_coat = find_text(kegg_hit, 'portal')
    kegg_baseplate = find_text(kegg_hit, 'baseplate')
    kegg_head = find_text(kegg_hit, 'head')
    kegg_tail = find_text(kegg_hit, 'tail')
    kegg_tail_fiber = find_text(kegg_hit, 'tail fiber')
    kegg_terminase = find_text(kegg_hit, 'terminase')
    kegg_integrase = find_text(kegg_hit, 'integrase')
    kegg_transposase = find_text(kegg_hit, 'transposase')
    
    viral_capsid = find_text(viral_hit, 'capsid')
    viral_portal = find_text(viral_hit, 'portal')
    viral_coat = find_text(viral_hit, 'portal')
    viral_baseplate = find_text(viral_hit, 'baseplate')
    viral_head = find_text(viral_hit, 'head')
    viral_tail = find_text(viral_hit, 'tail')
    viral_tail_fiber = find_text(viral_hit, 'tail fiber')
    viral_terminase = find_text(viral_hit, 'terminase')
    viral_integrase = find_text(viral_hit, 'integrase')
    viral_transposase = find_text(viral_hit, 'transposase')
    
    pfam_capsid = find_text(pfam_hit, 'capsid')
    pfam_portal = find_text(pfam_hit, 'portal')
    pfam_coat = find_text(pfam_hit, 'portal')
    pfam_baseplate = find_text(pfam_hit, 'baseplate')
    pfam_head = find_text(pfam_hit, 'head')
    pfam_tail = find_text(pfam_hit, 'tail')
    pfam_tail_fiber = find_text(pfam_hit, 'tail fiber')
    pfam_terminase = find_text(pfam_hit, 'terminase')
    pfam_integrase = find_text(pfam_hit, 'integrase')
    pfam_transposase = find_text(pfam_hit, 'transposase')
    return kegg_capsid, kegg_portal, kegg_coat, kegg_baseplate, kegg_head, kegg_tail, kegg_tail_fiber, kegg_terminase, kegg_integrase, kegg_transposase, viral_capsid, viral_portal, viral_coat, viral_baseplate, viral_head, viral_tail, viral_tail_fiber, viral_terminase, viral_integrase, viral_transposase, pfam_capsid, pfam_portal, pfam_coat, pfam_baseplate, pfam_head, pfam_tail, pfam_tail_fiber, pfam_terminase, pfam_integrase, pfam_transposase
        
    
    

dramv_amg_for_manual_curation_final = pd.DataFrame(list(dramv_amg_for_manual_curation_reformat_new.groupby('fasta').apply(get_hallmark)), index = group, columns = ['kegg_capsid', 'kegg_portal', 'kegg_coat', 'kegg_baseplate', 'kegg_head', 'kegg_tail', 'kegg_tail_fiber', 'kegg_terminase', 'kegg_integrase', 'kegg_transposase', 'viral_capsid', 'viral_portal', 'viral_coat', 'viral_baseplate', 'viral_head', 'viral_tail', 'viral_tail_fiber', 'viral_terminase', 'viral_integrase', 'viral_transposase', 'pfam_capsid', 'pfam_portal', 'pfam_coat', 'pfam_baseplate', 'pfam_head', 'pfam_tail', 'pfam_tail_fiber', 'pfam_terminase', 'pfam_integrase', 'pfam_transposase'])


dramv_amg_for_manual_curation_final.reset_index(inplace = True)
dramv_amg_for_manual_curation_final.rename(columns = {'index': 'contig'}, inplace = True)

amg_manual_curation = pd.merge(dramv_amg_for_manual_curation_final, checkv_for_amg_curation, left_on = 'contig', right_on = 'contig_id').drop('contig_id', axis = 1)

def remove_bracket(x):
    if str(x).startswith(r'['):
        return re.sub(r"[\[\]']", '', x)
    else:
        return x
    
            
amg_manual_curation = amg_manual_curation.applymap(remove_bracket)

amg_manual_curation['all_structure'] = amg_manual_curation['kegg_capsid'] + amg_manual_curation['kegg_portal'] + amg_manual_curation['kegg_coat'] + amg_manual_curation['kegg_baseplate'] + amg_manual_curation['kegg_head'] + amg_manual_curation['kegg_tail'] + amg_manual_curation['kegg_tail_fiber'] + amg_manual_curation['kegg_terminase'] +  amg_manual_curation['viral_capsid'] + amg_manual_curation['viral_portal'] + amg_manual_curation['viral_coat'] + amg_manual_curation['viral_baseplate'] + amg_manual_curation['viral_head'] + amg_manual_curation['viral_tail'] + amg_manual_curation['viral_tail_fiber'] + amg_manual_curation['viral_terminase'] + amg_manual_curation['pfam_capsid'] + amg_manual_curation['pfam_portal'] + amg_manual_curation['pfam_coat'] + amg_manual_curation['pfam_baseplate'] + amg_manual_curation['pfam_head'] + amg_manual_curation['pfam_tail'] + amg_manual_curation['pfam_tail_fiber'] + amg_manual_curation['pfam_terminase'] 

amg_manual_curation['structure_no_tail'] = amg_manual_curation['kegg_capsid'] + amg_manual_curation['kegg_portal'] + amg_manual_curation['kegg_coat'] + amg_manual_curation['kegg_baseplate'] + amg_manual_curation['kegg_head'] + amg_manual_curation['kegg_terminase'] +  amg_manual_curation['viral_capsid'] + amg_manual_curation['viral_portal'] + amg_manual_curation['viral_coat'] + amg_manual_curation['viral_baseplate'] + amg_manual_curation['viral_head'] + amg_manual_curation['viral_terminase'] + amg_manual_curation['pfam_capsid'] + amg_manual_curation['pfam_portal'] + amg_manual_curation['pfam_coat'] + amg_manual_curation['pfam_baseplate'] + amg_manual_curation['pfam_head']  + amg_manual_curation['pfam_terminase'] 

amg_manual_curation['integrase_or_transposase'] = amg_manual_curation['kegg_integrase'] + amg_manual_curation['kegg_transposase'] + amg_manual_curation['viral_integrase'] + amg_manual_curation['viral_transposase'] + amg_manual_curation['pfam_integrase'] + amg_manual_curation['pfam_transposase'] 

#amg_manual_curation.to_csv('/fs/ess/PAS0439/MING/virome/AMR_AMG_containing_contigs_manual_check/amg_manual_curation.csv', index = None)

os.chdir('/fs/scratch/PAS0439/Ming/results/amr_results/dramv')

dramv_amr = glob.glob('**/annotations.tsv', recursive = True)
dramv_amr_df = []

for file in dramv_amr:
    df = pd.read_csv(file, sep = '\t')
    dramv_amr_df.append(df)
    dramv_amr_for_manual_curation = pd.concat(dramv_amr_df)

dramv_amr_for_manual_curation.rename(columns = {'Unnamed: 0': 'gene_name'}, inplace = True)
dramv_amr_for_manual_curation.fasta = dramv_amr_for_manual_curation.fasta.apply(lambda x: x.replace('__full-cat_1', '||full'))
dramv_amr_for_manual_curation.fasta = dramv_amr_for_manual_curation.fasta.apply(lambda x: x.replace('__full-cat_2', '||full'))

dramv_amr_for_manual_curation_reformat = dramv_amr_for_manual_curation.get(['gene_name', 'fasta', 'kegg_hit', 'viral_hit', 'pfam_hits'])
dramv_amr_for_manual_curation_reformat_new = dramv_amr_for_manual_curation_reformat[dramv_amr_for_manual_curation_reformat.fasta.isin(amr_contigs_for_checking)]

group = dramv_amr_for_manual_curation_reformat_new.groupby('fasta').groups.keys()

group = dramv_amr_for_manual_curation_reformat_new.groupby('fasta').groups.keys()
dramv_amr_for_manual_curation_final = pd.DataFrame(list(dramv_amr_for_manual_curation_reformat_new.groupby('fasta').apply(get_hallmark)), index = group, columns = ['kegg_capsid', 'kegg_portal', 'kegg_coat', 'kegg_baseplate', 'kegg_head', 'kegg_tail', 'kegg_tail_fiber', 'kegg_terminase', 'kegg_integrase', 'kegg_transposase', 'viral_capsid', 'viral_portal', 'viral_coat', 'viral_baseplate', 'viral_head', 'viral_tail', 'viral_tail_fiber', 'viral_terminase', 'viral_integrase', 'viral_transposase', 'pfam_capsid', 'pfam_portal', 'pfam_coat', 'pfam_baseplate', 'pfam_head', 'pfam_tail', 'pfam_tail_fiber', 'pfam_terminase', 'pfam_integrase', 'pfam_transposase'])

dramv_amr_for_manual_curation_final.reset_index(inplace = True)
dramv_amr_for_manual_curation_final.rename(columns = {'index': 'contig'}, inplace = True)

amr_manual_curation = pd.merge(dramv_amr_for_manual_curation_final, checkv_for_amr_curation, left_on = 'contig', right_on = 'contig_id').drop('contig_id', axis = 1)

amr_manual_curation = amr_manual_curation.applymap(remove_bracket)

amr_manual_curation['all_structure'] = amr_manual_curation['kegg_capsid'] + amr_manual_curation['kegg_portal'] + amr_manual_curation['kegg_coat'] + amr_manual_curation['kegg_baseplate'] + amr_manual_curation['kegg_head'] + amr_manual_curation['kegg_tail'] + amr_manual_curation['kegg_tail_fiber'] + amr_manual_curation['kegg_terminase'] +  amr_manual_curation['viral_capsid'] + amr_manual_curation['viral_portal'] + amr_manual_curation['viral_coat'] + amr_manual_curation['viral_baseplate'] + amr_manual_curation['viral_head'] + amr_manual_curation['viral_tail'] + amr_manual_curation['viral_tail_fiber'] + amr_manual_curation['viral_terminase'] + amr_manual_curation['pfam_capsid'] + amr_manual_curation['pfam_portal'] + amr_manual_curation['pfam_coat'] + amr_manual_curation['pfam_baseplate'] + amr_manual_curation['pfam_head'] + amr_manual_curation['pfam_tail'] + amr_manual_curation['pfam_tail_fiber'] + amr_manual_curation['pfam_terminase'] 
amr_manual_curation['structure_no_tail'] = amr_manual_curation['kegg_capsid'] + amr_manual_curation['kegg_portal'] + amr_manual_curation['kegg_coat'] + amr_manual_curation['kegg_baseplate'] + amr_manual_curation['kegg_head'] + amr_manual_curation['kegg_terminase'] +  amr_manual_curation['viral_capsid'] + amr_manual_curation['viral_portal'] + amr_manual_curation['viral_coat'] + amr_manual_curation['viral_baseplate'] + amr_manual_curation['viral_head'] + amr_manual_curation['viral_terminase'] + amr_manual_curation['pfam_capsid'] + amr_manual_curation['pfam_portal'] + amr_manual_curation['pfam_coat'] + amr_manual_curation['pfam_baseplate'] + amr_manual_curation['pfam_head']  + amr_manual_curation['pfam_terminase'] 
amr_manual_curation['integrase_or_transposase'] = amr_manual_curation['kegg_integrase'] + amr_manual_curation['kegg_transposase'] + amr_manual_curation['viral_integrase'] + amr_manual_curation['viral_transposase'] + amr_manual_curation['pfam_integrase'] + amr_manual_curation['pfam_transposase'] 

amr_manual_curation.query('contig_length > 10000', inplace = True)

amr_manual_curation.to_csv('/fs/ess/PAS0439/MING/virome/AMR_AMG_containing_contigs_manual_check/amr_manual_curation.csv', index = None)
