==> vmags_check_helper/02_filter_based_on_checkv.ipynb <==
import os
import pandas as pd
import numpy as np
import glob

os.chdir('/fs/ess/PAS0439/MING/virome/checkv_res')

checkv_res = glob.glob("**/quality_summary.tsv", recursive=True)

df = []
for res in checkv_res:
    mydf = pd.read_csv(res, sep = '\t')
    df.append(mydf)

checkv_df = pd.concat(df)
checkv_df

checkv_filter = checkv_df.query('((contig_length > 10000) &( provirus == "No")|(proviral_length > 10000) &( provirus == "Yes"))').query('checkv_quality == "Complete"')#['contig_id']

checkv_filter

amg_check_df = pd.read_csv('/fs/ess/PAS0439/MING/virome/AMR_AMG_containing_contigs_manual_check/amg_manual_curation.csv')

amg_contigs = amg_check_df['contig']
amg_contigs_rename = [f.split('||')[0] for f in amg_contigs]

len(amg_contigs_rename)

def findoverlap(x):
    rows = []
    for f in x:
        row = checkv_filter[checkv_filter.contig_id.str.contains(f)]
        rows.append(row)
    df = pd.concat(rows)
    return df
        

checkv_res_for_amg = findoverlap(amg_contigs_rename)

checkv_res_for_amg.sort_values('contig_length', ascending = False)
checkv_res_for_amg
pd.set_option('display.max_rows', 2000)

checkv_res_for_amg.sort_values('contig_length', ascending = False)

checkv_filter

checkv_df[checkv_df.contig_id.str.contains('ERR2027900_249634')]

checkv_df.query('provirus == "Yes"').query('checkv_quality == "Complete"').sort_values('contig_length', ascending = False)

checkv_df.query('checkv_quality == "Complete"')
