==> vmags_check_helper/03_manual_check_helper.ipynb <==
import os
import pandas as pd 
import numpy as np 
import sys
import glob
from Bio import SeqIO
import re

os.chdir('/fs/ess/PAS0439/MING/virome/amg_analysis/dram_annotation')

dram_res = glob.glob('**/annotations.tsv' ,recursive= True)

dram_df = []
for dram in dram_res:
    df = pd.read_csv(dram, sep = '\t')
    dram_df.append(df)

dram_res = pd.concat(dram_df)[['scaffold', 'gene_position', 'kegg_hit', 'cazy_hits', 'viral_hit', 'pfam_hits', 'vogdb_hit', 'amg_flags']]

checkv_contigs = list(np.loadtxt('amg_contigs.txt' , dtype = 'str'))

checkv_contigs_renamed = [f.replace('>', '').replace('||', '__') for f in checkv_res]

def rename(x):
    renamed = re.sub('-cat.*', '', x)
    return renamed
dram_res.scaffold = dram_res.scaffold.apply(rename)

dram_res_manual_check = dram_res[dram_res.scaffold.isin(checkv_contigs_renamed)]

check_list = list(dram_res_manual_check['scaffold'])

checkv_res = pd.read_csv('checkv_genomic_island_checkv_res/quality_summary.tsv', sep ='\t')[['contig_id', 'contig_length']]
checkv_res.contig_id = checkv_res.contig_id.apply(lambda x: x.replace('||', '__'))

individual_check = pd.merge(checkv_res, dram_res_manual_check, left_on = 'contig_id', right_on = 'scaffold')

individual_check.drop('scaffold', inplace = True)

dram_res_manual_check.to_csv('amg_contig_individual_check.tsv', sep = '\t', index = None)

check_list = list(set(individual_check['contig_id']))
check_list[0]
individual_check.drop('scaffold',axis = 1,  inplace = True)

individual_check.to_csv('amg_contig_individual_check.tsv', sep = '\t', index = None)
